test cuda13.0 build based on pytorch:25.08-py3

+++ superbench/benchmarks/micro_benchmarks/gpu_stream/gpu_stream_kernels.cu
@@ -85,7 +85,7 @@ __global__ void CopyKernel(double *tgt, const double *src) {
  * @param[in] src The source array containing the data to be scaled.
  * @param[in] scalar The scalar value used to scale the source data.
  */
-__global__ void ScaleKernel(double *tgt, const double *src, const long scalar) {
+__global__ void ScaleKernel(double *tgt, const double *src, const double scalar) {
     uint64_t index = blockIdx.x * blockDim.x * kNumLoopUnrollAlias + threadIdx.x;
     double val[kNumLoopUnrollAlias];
 #pragma unroll
